re added OpImage changes

+++ test/spv_test.cpp
@@ -77,7 +77,7 @@ namespace Sol {
         TEST_EQ("ubo_type", ubo_parent_name, static_cast<uint32_t>(Spv::Name::VAR), skip);
 
         Spv::Ptr *ubo_ptr = get_T<Spv::Ptr>(spv.types, ubo.ptr_id);
-        int ubo_ptr_id = spv.types[1].id;
+        uint32_t ubo_ptr_id = spv.types[1].id;
         Spv::Type *ubo_ptr_parent = get_parent(spv.types, ubo.ptr_id);
         ASSERT(ubo_ptr_parent != nullptr, "ID not found");
         TEST_EQ("ubo_ptr_id", ubo.ptr_id, ubo_ptr_parent->id, skip);
@@ -89,14 +89,14 @@ namespace Sol {
         uint32_t in_color_id = spv.types[5].id;
         uint32_t deco_flags_location = static_cast<uint32_t>(Flags::LOCATION);
         TEST_EQ("in_color_flags", in_color.deco_info.flags, deco_flags_location, skip);
-        TEST_EQ("in_color_location", in_color.deco_info.location, 1, skip);
+        TEST_EQ("in_color_location", in_color.deco_info.location, static_cast<uint32_t>(1), skip);
         Spv::Type *in_color_parent = get_parent(spv.types, in_color_id);
         TEST_EQ("in_color_type", static_cast<uint32_t>(in_color_parent->name), static_cast<uint32_t>(Spv::Name::VAR), skip);
 
         Spv::Var tex_sampler = *(reinterpret_cast<Spv::Var*>(spv.types[6].data));
         TEST_EQ("tex_sampler_flags", tex_sampler.deco_info.flags, deco_flags_binding | deco_flags_desc_set, skip);
-        TEST_EQ("tex_sampler_binding", tex_sampler.deco_info.binding, 1, skip);
-        TEST_EQ("tex_sampler_desc_set", tex_sampler.deco_info.desc_set, 1, skip);
+        TEST_EQ("tex_sampler_binding", tex_sampler.deco_info.binding, static_cast<uint32_t>(1), skip);
+        TEST_EQ("tex_sampler_desc_set", tex_sampler.deco_info.desc_set, static_cast<uint32_t>(1), skip);
         Spv::Type *tex_sampler_parent = get_parent(spv.types, in_color_id);
         TEST_EQ("tex_sampler_type", static_cast<uint32_t>(tex_sampler_parent->name), static_cast<uint32_t>(Spv::Name::VAR), skip);
 
@@ -104,35 +104,35 @@ namespace Sol {
         Spv::Float floating_point = *reinterpret_cast<Spv::Float*>(types[7].data);
         Spv::Int unsigned_int = *reinterpret_cast<Spv::Int*>(types[9].data);
         Spv::Int signed_int = *reinterpret_cast<Spv::Int*>(types[12].data);
-        TEST_EQ("floating_point_id", types[7].id, 6, skip);
+        TEST_EQ("floating_point_id", types[7].id, static_cast<uint32_t>(6), skip);
         TEST_EQ("floating_point_type", static_cast<uint32_t>(types[7].name), static_cast<uint32_t>(Spv::Name::FLOAT), skip);
-        TEST_EQ("floating_point_width", floating_point.width, 32, skip);
-        TEST_EQ("unsigned_int_id", types[9].id, 8, skip);
+        TEST_EQ("floating_point_width", floating_point.width, static_cast<uint32_t>(32), skip);
+        TEST_EQ("unsigned_int_id", types[9].id, static_cast<uint32_t>(8), skip);
         TEST_EQ("unsigned_int_type", static_cast<uint32_t>(types[9].name), static_cast<uint32_t>(Spv::Name::INT), skip);
         TEST_EQ("unsigned_int_sign", unsigned_int.sign, false, skip);
-        TEST_EQ("signed_int_width", unsigned_int.width, 32, skip);
-        TEST_EQ("signed_int_id", types[12].id, 14, skip);
+        TEST_EQ("signed_int_width", unsigned_int.width, static_cast<uint32_t>(32), skip);
+        TEST_EQ("signed_int_id", types[12].id, static_cast<uint32_t>(14), skip);
         TEST_EQ("signed_int_type", static_cast<uint32_t>(types[12].name), static_cast<uint32_t>(Spv::Name::INT), skip);
         TEST_EQ("signed_int_sign", signed_int.sign, true, skip);
-        TEST_EQ("signed_int_width", signed_int.width, 32, skip);
+        TEST_EQ("signed_int_width", signed_int.width, static_cast<uint32_t>(32), skip);
 
         // VEC
         Spv::Vector vec4_float = *reinterpret_cast<Spv::Vector*>(types[8].data);
         Spv::Vector vec2_float = *reinterpret_cast<Spv::Vector*>(types[16].data);
         Spv::Vector vec3_float = *reinterpret_cast<Spv::Vector*>(types[19].data);
-        int float_id = types[7].id;
+        uint32_t float_id = types[7].id;
         TEST_EQ("vec4_type_id", vec4_float.type_id, float_id, skip);
-        TEST_EQ("vec4_type_length", vec4_float.length, 4, skip);
+        TEST_EQ("vec4_type_length", vec4_float.length, static_cast<uint32_t>(4), skip);
         TEST_EQ("vec2_type_id", vec2_float.type_id, float_id, skip);
-        TEST_EQ("vec2_type_length", vec2_float.length, 2, skip);
+        TEST_EQ("vec2_type_length", vec2_float.length, static_cast<uint32_t>(2), skip);
         TEST_EQ("vec3_type_id", vec3_float.type_id, float_id, skip);
-        TEST_EQ("vec3_type_length", vec3_float.length, 3, skip);
+        TEST_EQ("vec3_type_length", vec3_float.length, static_cast<uint32_t>(3), skip);
 
         // MATRIX
         Spv::Matrix mat4_v4_float = *reinterpret_cast<Spv::Matrix*>(types[13].data);
-        int vec4_float_id = types[8].id;
+        uint32_t vec4_float_id = types[8].id;
         TEST_EQ("mat4_type_id", mat4_v4_float.type_id, vec4_float_id, skip);
-        TEST_EQ("mat4_col_count", mat4_v4_float.column_count, 4, skip);
+        TEST_EQ("mat4_col_count", mat4_v4_float.column_count, static_cast<uint32_t>(4), skip);
 
         spv.kill();
         mem_free(pcode, HEAP);
